Don't archive U-Boot source
Rely on Jenkins to place a copy of the U-Boot source tree into the test
job. This limits us to a single copy of the tree per job definition,
rather than archiving a copy each time the job is run. test.sh ensures
the build and test run using the same commit by explicitly checking out
the commit that build.sh used.

Signed-off-by: Stephen Warren <swarren@nvidia.com>