Use Linaro gcc-7.2.1-2017.11 for compilation
Recent versions of U-Boot won't compile with anything older than gcc-6.
No LTS version of Ubuntu packages a new enough compiler:-( Update the CI
scripts to rely on a manually installed version of the Linaro toolchain.
Hopefully this is compatible with all the old branches of U-Boot that we
build too...

Signed-off-by: Stephen Warren <swarren@nvidia.com>