Extend CONFIG_BOOTDELAY
CONFIG_BOOTDELAY is 2s by default. This is little enough time that it
can take longer than this for the u-boot-reset script to exit after
having started the board booting and for test/py to process the console
output and interrupt the boot process. This causes intermittent test
failures. Extend the timeout to prevent this from happening. Autoboot
isn't completely disabled to ensure that the autoboot interruption
feature is still tested.

Signed-off-by: Stephen Warren <swarren@nvidia.com>