Remove dfu-util patches
We still need a custom build of dfu-util since Ubuntu Xenial's packaged
dfu-util doesn't actually implement the -p command-line option despite
allowing it to be specified!

However, it looks like we can get rid of the custom patch to the timeout;
at least multiple upstream versions of U-Boot don't need it. Checking in
this version to determine whether any downstream L4T branches still need
the patch or not.

Signed-off-by: Stephen Warren <swarren@nvidia.com>